+++ hangman_server.c
@@ -14,8 +14,8 @@ int numClients = 0;
 pthread_mutex_t mutex;
 
 struct serverMessage {
-    int message_flag;
-    int word_length;
+    short message_flag;
+    short word_length;
     int num_incorrect;
     char data[256];
 };
